pyproject.toml
[project]
name = "rp2040-modbus-poll"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymodbus>=3.12.1",
    "pyserial>=3.5",
]
